update schema, split upstash into own files
