+++ apps/web/app/proxy/[domain]/[key]/page.tsx
@@ -30,8 +30,7 @@ export async function generateMetadata({
     title: unescape(data.title || ""),
     description: unescape(data.description || ""),
     image: data.image,
-    // TODO replace this with value from DB
-    video: "https://static.mux.com/videos/sandwich.mp4",
+    video: data.video,
     icons: `${GOOGLE_FAVICON_URL}${apexDomain}`,
     noIndex: true,
   });
