+++ apps/web/lib/upstash/format-redis-link.ts
@@ -1,0 +1,44 @@
+import { isIframeable } from "@dub/utils";
+import { LinkProps, RedisLinkProps } from "../types";
+
+export async function formatRedisLink(
+  link: LinkProps,
+): Promise<RedisLinkProps> {
+  const {
+    id,
+    domain,
+    url,
+    trackConversion,
+    password,
+    proxy,
+    rewrite,
+    expiresAt,
+    expiredUrl,
+    ios,
+    android,
+    geo,
+    doIndex,
+    projectId,
+  } = link;
+  const hasPassword = password && password.length > 0 ? true : false;
+
+  return {
+    id,
+    ...(url && { url }), // on free plans you cannot set a root domain redirect, hence URL is undefined
+    ...(trackConversion && { trackConversion: true }),
+    ...(hasPassword && { password: true }),
+    ...(proxy && { proxy: true }),
+    ...(url &&
+      rewrite && {
+        rewrite: true,
+        iframeable: await isIframeable({ url, requestDomain: domain }),
+      }),
+    ...(expiresAt && { expiresAt: new Date(expiresAt) }),
+    ...(expiredUrl && { expiredUrl }),
+    ...(ios && { ios }),
+    ...(android && { android }),
+    ...(geo && { geo: geo as object }),
+    ...(projectId && { projectId }), // projectId can be undefined for anonymous links
+    ...(doIndex && { doIndex: true }),
+  };
+}
