+++ apps/web/lib/upstash/index.ts
@@ -1,0 +1,10 @@
+export * from "./format-redis-link";
+export * from "./ratelimit";
+export * from "./record-metatags";
+import { Redis } from "@upstash/redis";
+
+// Initiate Redis instance by connecting to REST URL
+export const redis = new Redis({
+  url: process.env.UPSTASH_REDIS_REST_URL || "",
+  token: process.env.UPSTASH_REDIS_REST_TOKEN || "",
+});
