+++ apps/web/lib/upstash/ratelimit.ts
@@ -1,0 +1,31 @@
+import { Ratelimit } from "@upstash/ratelimit";
+import { Redis } from "@upstash/redis";
+
+export const ratelimitRedis = new Redis({
+  url:
+    process.env.RATELIMIT_UPSTASH_REDIS_REST_URL ||
+    process.env.UPSTASH_REDIS_REST_URL ||
+    "",
+  token:
+    process.env.RATELIMIT_UPSTASH_REDIS_REST_TOKEN ||
+    process.env.UPSTASH_REDIS_REST_TOKEN ||
+    "",
+});
+
+// Create a new ratelimiter, that allows 10 requests per 10 seconds by default
+export const ratelimit = (
+  requests: number = 10,
+  seconds:
+    | `${number} ms`
+    | `${number} s`
+    | `${number} m`
+    | `${number} h`
+    | `${number} d` = "10 s",
+) => {
+  return new Ratelimit({
+    redis: ratelimitRedis,
+    limiter: Ratelimit.slidingWindow(requests, seconds),
+    analytics: true,
+    prefix: "dub",
+  });
+};
